Refactoring SERVER
