+++ Server/Models/Entities/PersonaEntity.cs
@@ -1,15 +1,11 @@
-﻿namespace DemoBlazorApp.Server.Entities
+﻿namespace DemoBlazorApp.Server.Models.Entities
 {
     public partial class PersonaEntity
     {
         public int PersonaId { get; set; }
-
         public string Cognome { get; set; }
-
         public string Nome { get; set; }
-
         public string Email { get; set; }
-
         public string Telefono { get; set; }
     }
 }
